pyproject.toml
[tool.poetry]
name = "ispee"
version = "0.1.0"
description = ""
authors = ["Tim Martin <tim@timmart.in>"]

[tool.poetry.dependencies]
python = "^3.10"
prometheus-client = "0.15.0"
rich = "12.6.0"
click = "8.1.3"
PyYAML = "^6.0"
attrs = "22.1.0"
httpx = "0.23.0"
yarl = "1.8.1"
icmplib = "^3.0.3"
anyio = "^3.6.2"
black = "22.10.0"
isort = "5.10.1"
types-pyyaml = "6.0.12"
dnspython = "^2.2.1"

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"
mypy = "^0.931"
flake8 = "^4.0.1"
isort = "^5.10.1"
types-PyYAML = "^6.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.mypy]
files = ["src", "tests"]
check_untyped_defs = true
follow_imports_for_stubs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
implicit_reexport = false
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true
# disallow_any_unimported = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "icmplib",
    "trio_util"
]
ignore_missing_imports = true
